PAE-220 push view handler registering code up into app/routes.js
use same pattern for adding GET handler to index (/) page
